Summary and comparison of the levels in categorical columns

For a single dataframe, summarise the levels of each categorical column. If two dataframes are supplied, compare the levels of categorical features that appear in both dataframes. For grouped dataframes, summarise the levels of categorical features separately for each group.

inspect_cat(df1, df2 = NULL, show_plot = FALSE)

Arguments

df1
A dataframe.

df2
An optional second data frame for comparing categorical levels. Defaults to NULL.

show_plot
(Deprecated) Logical flag indicating whether a plot should be shown. Superseded by the function show_plot() and will be dropped in a future version.

Value

A tibble summarising or comparing the categorical features in one or a pair of dataframes.

Details

For a single dataframe, the tibble returned contains the columns:

col_name, character vector containing column names of df1.

cnt integer column containing count of unique levels found in each column, including NA.

common, a character column containing the name of the most common level.

common_pcnt, the percentage of each column occupied by the most common level shown in common.

levels, a named list containing relative frequency tibbles for each feature.

For a pair of dataframes, the tibble returned contains the columns:

col_name, character vector containing names of columns appearing in both df1 and df2.

jsd, a numeric column containing the Jensen-Shannon divergence. This measures the difference in relative frequencies of levels in a pair of categorical features. Values near to 0 indicate agreement of the distributions, while 1 indicates disagreement.

fisher_p, the p-value corresponding to Fisher's exact test. A small p indicates evidence that the the two sets of relative frequencies are actually different.

lvls_1, lvls_2, the relative frequency of levels in each of df1 and df2.

For a grouped dataframe, the tibble returned is as for a single dataframe, but where the first k columns are the grouping columns. There will be as many rows in the result as there are unique combinations of the grouping variables.

See also

Examples

# Load dplyr for starwars data & pipe
library(dplyr)

#> 
#> Attaching package: ‘dplyr’

#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag

#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

# Single dataframe summary
inspect_cat(starwars)

#> # A tibble: 7 x 5
#>   col_name     cnt common common_pcnt levels           
#>   <chr>      <int> <chr>        <dbl> <named list>     
#> 1 eye_color     15 brown        24.1  <tibble [15 × 3]>
#> 2 gender         5 male         71.3  <tibble [5 × 3]> 
#> 3 hair_color    13 none         42.5  <tibble [13 × 3]>
#> 4 homeworld     49 Naboo        12.6  <tibble [49 × 3]>
#> 5 name          87 Ackbar        1.15 <tibble [87 × 3]>
#> 6 skin_color    31 fair         19.5  <tibble [31 × 3]>
#> 7 species       38 Human        40.2  <tibble [38 × 3]>

# Paired dataframe comparison
inspect_cat(starwars, starwars[1:20, ])

#> # A tibble: 7 x 5
#>   col_name      jsd fisher_p lvls_1            lvls_2           
#>   <chr>       <dbl>    <dbl> <named list>      <named list>     
#> 1 eye_color  0.0936 0.750    <tibble [15 × 3]> <tibble [8 × 3]> 
#> 2 gender     0.0653 0.0819   <tibble [5 × 3]>  <tibble [4 × 3]> 
#> 3 hair_color 0.261  0.000843 <tibble [13 × 3]> <tibble [10 × 3]>
#> 4 homeworld  0.394  0.359    <tibble [49 × 3]> <tibble [11 × 3]>
#> 5 name       0.573  1.000    <tibble [87 × 3]> <tibble [20 × 3]>
#> 6 skin_color 0.288  0.990    <tibble [31 × 3]> <tibble [10 × 3]>
#> 7 species    0.310  0.810    <tibble [38 × 3]> <tibble [6 × 3]> 

# Grouped dataframe summary
starwars %>% group_by(gender) %>% inspect_cat()

#> # A tibble: 30 x 6
#>    gender        col_name     cnt common   common_pcnt levels           
#>    <chr>         <chr>      <int> <chr>          <dbl> <list>           
#>  1 female        eye_color     13 brown          25.8  <tibble [13 × 3]>
#>  2 female        hair_color    10 none           46.8  <tibble [10 × 3]>
#>  3 female        homeworld     43 Naboo          11.3  <tibble [43 × 3]>
#>  4 female        name          62 Ackbar          1.61 <tibble [62 × 3]>
#>  5 female        skin_color    22 fair           22.6  <tibble [22 × 3]>
#>  6 female        species       32 Human          41.9  <tibble [32 × 3]>
#>  7 hermaphrodite eye_color      2 red            66.7  <tibble [2 × 3]> 
#>  8 hermaphrodite hair_color     1 NA            100    <tibble [1 × 3]> 
#>  9 hermaphrodite homeworld      2 Tatooine       66.7  <tibble [2 × 3]> 
#> 10 hermaphrodite name           3 C-3PO          33.3  <tibble [3 × 3]> 
#> # … with 20 more rows