Summarise and compare the levels within each categorical feature in one or two dataframes.

Summarise and compare the levels within each categorical feature in one or two dataframes.

inspect_cat(df1, df2 = NULL, show_plot = FALSE)

Arguments

df1
A dataframe

df2
An optional second data frame for comparing categorical levels. Defaults to NULL.

show_plot
(Deprecated) Logical flag indicating whether a plot should be shown. Superseded by the function show_plot() and will be dropped in a future version.

Value

A tibble summarising and comparing the categorical features in one or a pair of data frames.

Details

When only df1 is specified, a tibble is returned which contains summaries of the categorical levels in df1.

col_name character vector containing column names of df1.

cnt integer column containing count of unique levels found in each column (including NA).

common character column containing the name of the most common level.

common_pcnt percentage of each column occupied by the most common level shown in common.

levels relative frequency of levels in each column.

When both df1 and df2 are specified, the relative frequencies of levels across columns common both dataframes are compared. In particular, the population stability index, and Fisher's exact test are performed as part of the comparison.

col_name character vector containing column names of df1 and df2.

jsd numeric column containing the Jensen-Shannon divergence. This measures the difference in distribution of a pair of categorical features. Values near to 0 indicate agreement of the distributions, while 1 indicates disagreement.

fisher_p p-value corresponding to Fisher's exact test. A small p indicates evidence that the the two sets of relative frequencies are actually different.

lvls_1, lvls_2 relative frequency of levels in each of df1 and df2.

Examples

data("starwars", package = "dplyr")
inspect_cat(starwars)

#> # A tibble: 7 x 5
#>   col_name     cnt common common_pcnt levels           
#>   <chr>      <int> <chr>        <dbl> <named list>     
#> 1 eye_color     15 brown        24.1  <tibble [15 × 3]>
#> 2 gender         5 male         71.3  <tibble [5 × 3]> 
#> 3 hair_color    13 none         42.5  <tibble [13 × 3]>
#> 4 homeworld     49 Naboo        12.6  <tibble [49 × 3]>
#> 5 name          87 Ackbar        1.15 <tibble [87 × 3]>
#> 6 skin_color    31 fair         19.5  <tibble [31 × 3]>
#> 7 species       38 Human        40.2  <tibble [38 × 3]>

# compare the levels in two data frames
inspect_cat(starwars, starwars[1:20, ])

#> # A tibble: 7 x 5
#>   col_name      jsd fisher_p lvls_1            lvls_2           
#>   <chr>       <dbl>    <dbl> <named list>      <named list>     
#> 1 eye_color  0.0936 0.750    <tibble [15 × 3]> <tibble [8 × 3]> 
#> 2 gender     0.0653 0.0819   <tibble [5 × 3]>  <tibble [4 × 3]> 
#> 3 hair_color 0.261  0.000843 <tibble [13 × 3]> <tibble [10 × 3]>
#> 4 homeworld  0.394  0.359    <tibble [49 × 3]> <tibble [11 × 3]>
#> 5 name       0.573  1.000    <tibble [87 × 3]> <tibble [20 × 3]>
#> 6 skin_color 0.288  0.990    <tibble [31 × 3]> <tibble [10 × 3]>
#> 7 species    0.310  0.810    <tibble [38 × 3]> <tibble [6 × 3]>