Summarise and compare the memory usage in one or two dataframes.

Summarise and compare the memory usage in one or two dataframes.

inspect_mem(df1, df2 = NULL, show_plot = FALSE)

Arguments

df1
A data frame.

df2
An optional second data frame with which to comparing memory usage. Defaults to NULL.

show_plot
(Deprecated) Logical flag indicating whether a plot should be shown. Superseded by the function show_plot() and will be dropped in a future version.

Value

A tibble summarising and comparing the columnwise memory usage for one or a pair of data frames.

Details

When df1 is specified and df2 = NULL, a tibble summarising columnwise memory usage in descending order of size is returned:

col_name character vector containing column names of df1.

size character vector containing display-friendly memory usage of each column.

pcnt the percentage of the dataframe's total memory footprint used by each column.

When both df1 and df2 are specified, column memory usages are jointly tabulated for both data frames. Rows are sorted in descending order of size as they appear in df1:

col_name character vector containing column names of df1 and df2.

size_1, size_2 character vector containing memory usage of each column in each of df1 and df2.

pcnt_1, pcnt_2 the percentage of total memory usage of each column within each of df1 and df2.

Examples

data("starwars", package = "dplyr")
# get tibble of column memory usage for the starwars data
inspect_mem(starwars)

#> # A tibble: 13 x 3
#>    col_name   size        pcnt
#>    <chr>      <chr>      <dbl>
#>  1 films      19.54 Kb  36.5  
#>  2 starships  7.27 Kb   13.6  
#>  3 name       6.13 Kb   11.5  
#>  4 vehicles   5.8 Kb    10.8  
#>  5 homeworld  3.52 Kb    6.58 
#>  6 species    2.88 Kb    5.39 
#>  7 skin_color 2.59 Kb    4.85 
#>  8 eye_color  1.57 Kb    2.93 
#>  9 hair_color 1.41 Kb    2.63 
#> 10 gender     976 bytes  1.78 
#> 11 mass       744 bytes  1.36 
#> 12 birth_year 744 bytes  1.36 
#> 13 height     400 bytes  0.730

# compare memory usage 
inspect_mem(starwars, starwars[1:10, -3])

#> # A tibble: 13 x 5
#>    col_name   size_1    size_2    pcnt_1 pcnt_2
#>    <chr>      <chr>     <chr>      <dbl>  <dbl>
#>  1 films      19.54 Kb  4.05 Kb   36.5   40.9  
#>  2 starships  7.27 Kb   1.36 Kb   13.6   13.8  
#>  3 name       6.13 Kb   824 bytes 11.5    8.14 
#>  4 vehicles   5.8 Kb    976 bytes 10.8    9.64 
#>  5 homeworld  3.52 Kb   416 bytes  6.58   4.11 
#>  6 species    2.88 Kb   288 bytes  5.39   2.85 
#>  7 skin_color 2.59 Kb   528 bytes  4.85   5.22 
#>  8 eye_color  1.57 Kb   464 bytes  2.93   4.58 
#>  9 hair_color 1.41 Kb   528 bytes  2.63   5.22 
#> 10 gender     976 bytes 288 bytes  1.78   2.85 
#> 11 mass       744 bytes NA         1.36  NA    
#> 12 birth_year 744 bytes 176 bytes  1.36   1.74 
#> 13 height     400 bytes 96 bytes   0.730  0.949