Summary and comparison of memory usage of dataframe columns

For a single dataframe, summarise the memory usage in each column. If two dataframes are supplied, compare memory usage for columns appearing in both dataframes. For grouped dataframes, summarise the memory usage separately for each group.

inspect_mem(df1, df2 = NULL, show_plot = FALSE)

Arguments

df1
A data frame.

df2
An optional second data frame with which to comparing memory usage. Defaults to NULL.

show_plot
(Deprecated) Logical flag indicating whether a plot should be shown. Superseded by the function show_plot() and will be dropped in a future version.

Value

A tibble summarising and comparing the columnwise memory usage for one or a pair of data frames.

Details

For a single dataframe, the tibble returned contains the columns:

col_name, a character vector containing column names of df1.

size, a character vector containing display-friendly memory usage of each column.

pcnt, the percentage of the dataframe's total memory footprint used by each column.

For a pair of dataframes, the tibble returned contains the columns:

col_name, a character vector containing column names of df1 and df2.

size_1, size_2, a character vector containing memory usage of each column in each of df1 and df2.

pcnt_1, pcnt_2, the percentage of total memory usage of each column within each of df1 and df2.

For a grouped dataframe, the tibble returned is as for a single dataframe, but where the first k columns are the grouping columns. There will be as many rows in the result as there are unique combinations of the grouping variables.

See also

Examples

# Load dplyr for starwars data & pipe
library(dplyr)

# Single dataframe summary
inspect_mem(starwars)

#> # A tibble: 13 x 3
#>    col_name   size        pcnt
#>    <chr>      <chr>      <dbl>
#>  1 films      19.54 Kb  36.5  
#>  2 starships  7.27 Kb   13.6  
#>  3 name       6.13 Kb   11.5  
#>  4 vehicles   5.8 Kb    10.8  
#>  5 homeworld  3.52 Kb    6.58 
#>  6 species    2.88 Kb    5.39 
#>  7 skin_color 2.59 Kb    4.85 
#>  8 eye_color  1.57 Kb    2.93 
#>  9 hair_color 1.41 Kb    2.63 
#> 10 gender     976 bytes  1.78 
#> 11 mass       744 bytes  1.36 
#> 12 birth_year 744 bytes  1.36 
#> 13 height     400 bytes  0.730

# Paired dataframe comparison
inspect_mem(starwars, starwars[1:20, ])

#> # A tibble: 13 x 5
#>    col_name   size_1    size_2    pcnt_1 pcnt_2
#>    <chr>      <chr>     <chr>      <dbl>  <dbl>
#>  1 films      19.54 Kb  7.23 Kb   36.5   41.4  
#>  2 starships  7.27 Kb   2.58 Kb   13.6   14.7  
#>  3 name       6.13 Kb   1.49 Kb   11.5    8.53 
#>  4 vehicles   5.8 Kb    1.74 Kb   10.8    9.96 
#>  5 homeworld  3.52 Kb   816 bytes  6.58   4.56 
#>  6 species    2.88 Kb   552 bytes  5.39   3.08 
#>  7 skin_color 2.59 Kb   808 bytes  4.85   4.51 
#>  8 eye_color  1.57 Kb   664 bytes  2.93   3.71 
#>  9 hair_color 1.41 Kb   736 bytes  2.63   4.11 
#> 10 gender     976 bytes 384 bytes  1.78   2.14 
#> 11 mass       744 bytes 208 bytes  1.36   1.16 
#> 12 birth_year 744 bytes 208 bytes  1.36   1.16 
#> 13 height     400 bytes 176 bytes  0.730  0.983

# Grouped dataframe summary
starwars %>% group_by(gender) %>% inspect_mem()

#> # A tibble: 60 x 4
#>    gender col_name   size       pcnt
#>    <chr>  <chr>      <chr>     <dbl>
#>  1 female films      13.58 Kb  35.1 
#>  2 female starships  5.66 Kb   14.6 
#>  3 female name       4.42 Kb   11.4 
#>  4 female vehicles   4.27 Kb   11.0 
#>  5 female homeworld  2.98 Kb    7.71
#>  6 female species    2.34 Kb    6.03
#>  7 female skin_color 1.82 Kb    4.70
#>  8 female eye_color  1.26 Kb    3.25
#>  9 female hair_color 1.05 Kb    2.70
#> 10 female mass       544 bytes  1.37
#> # … with 50 more rows