Summarise and compare the rate of missingness in one or two dataframes.

Summarise and compare the rate of missingness in one or two dataframes.

inspect_na(df1, df2 = NULL, show_plot = FALSE)

Arguments

df1
A data frame

df2
An optional second data frame for making columnwise comparison of missingness. Defaults to NULL.

show_plot
(Deprecated) Logical flag indicating whether a plot should be shown. Superseded by the function show_plot() and will be dropped in a future version.

Value

A tibble summarising the count and percentage of columnwise missingness for one or a pair of data frames.

Details

When df1 is specified and df2 = NULL, a tibble containing columnwise summaries of missing values is returned, with columns:

col_name character vector containing column names of df1.

cnt integer vector containing the number of missing values by column.

pcnt the percentage of records in each columns that is missing.

When both df1 and df2 are specified, missingness is compared across all columns in both dataframes. A test of the null hypothesis that the rate of missingness is the same across the same column in either dataframe.

col_name the name of the columns occurring in either df1 or df2.

cnt_1, cnt_2 pair of integer vectors containing counts of missing entries for each column in df1 and df2.

pcnt_1, pcnt_2 pair of columns containing percentage of missing entries for each column in df1 and df2.

p_value p-value associated with test of equivalence of rates of missingness. Small values indicate evidence that the rate of missingness differs for a column occurring in both df1 and df2.

Examples

data("starwars", package = "dplyr")
# inspect missingness in starwars data
inspect_na(starwars)

#> # A tibble: 13 x 3
#>    col_name     cnt  pcnt
#>    <chr>      <int> <dbl>
#>  1 birth_year    44 50.6 
#>  2 mass          28 32.2 
#>  3 homeworld     10 11.5 
#>  4 height         6  6.90
#>  5 hair_color     5  5.75
#>  6 species        5  5.75
#>  7 gender         3  3.45
#>  8 name           0  0   
#>  9 skin_color     0  0   
#> 10 eye_color      0  0   
#> 11 films          0  0   
#> 12 vehicles       0  0   
#> 13 starships      0  0   

# compare two dataframes
inspect_na(starwars, starwars[1:30, ])

#> # A tibble: 13 x 6
#>    col_name   cnt_1 pcnt_1 cnt_2 pcnt_2   p_value
#>    <chr>      <int>  <dbl> <int>  <dbl>     <dbl>
#>  1 birth_year    44  50.6      4  13.3   0.000777
#>  2 mass          28  32.2      3  10     0.0328  
#>  3 homeworld     10  11.5      3  10     1.000   
#>  4 height         6   6.90     1   3.33  0.792   
#>  5 hair_color     5   5.75     5  16.7   0.143   
#>  6 species        5   5.75     0   0     0.413   
#>  7 gender         3   3.45     3  10     0.356   
#>  8 name           0   0        0   0    NA       
#>  9 skin_color     0   0        0   0    NA       
#> 10 eye_color      0   0        0   0    NA       
#> 11 films          0   0        0   0    NA       
#> 12 vehicles       0   0        0   0    NA       
#> 13 starships      0   0        0   0    NA