Summarise and compare the numeric variables within one or two dataframes

Summarise and compare the numeric variables within one or two dataframes

inspect_num(df1, df2 = NULL, breaks = 20, include_int = TRUE,
  show_plot = FALSE)

Arguments

df1
A dataframe.

df2
An optional second dataframe for comparing categorical levels. Defaults to NULL.

breaks
Integer number of breaks used for histogram bins, passed to graphics::hist(). Defaults to 20.

include_int
Logical flag, whether to include integer columns in numeric summaries. Defaults to TRUE.

show_plot
(Deprecated) Logical flag indicating whether a plot should be shown. Superseded by the function show_plot() and will be dropped in a future version. hist(..., breaks). See ?hist for more details.

Value

A tibble containing statistical summaries of the numeric columns of df1, or comparing the histograms of df1 and df2.

Details

If only df1 is specified, inspect_num() returns a tibble with columns

col_name, a character vector containing the column names in df1

min, q1, median, mean, q3, max and sd: the minimum, lower quartile, median, mean, upper quartile, maximum and standard deviation for each numeric column.

pcnt_na, the percentage of each numeric feature that is missing

hist, a named list of tibbles containing the relative frequency of values in a falling in bins determined by breaks.

If both df1 and df2 are specified, the tibble has columns

col_name character vector containing the column names in df1 and df2

hist_1, hist_2 list column for histograms of each of df1 and df2. Where a column appears in both dataframe, the bins used for df1 are reused to calculate histograms for df2.

jsd numeric column containing the Jensen-Shannon divergence. This measures the difference in distribution of a pair of binned numeric features. Values near to 0 indicate agreement of the distributions, while 1 indicates disagreement.

fisher_p p-value corresponding to Fisher's exact test. A small p indicates evidence that the two histograms are actually different.

Examples

data("starwars", package = "dplyr")
# show summary statistics for starwars
inspect_num(starwars)

#> # A tibble: 3 x 10
#>   col_name     min    q1 median  mean    q3   max    sd pcnt_na hist            
#>   <chr>      <dbl> <dbl>  <dbl> <dbl> <dbl> <dbl> <dbl>   <dbl> <named list>    
#> 1 height        66 167      180 174.  191     264  34.8    6.90 <tibble [23 × 2…
#> 2 mass          15  55.6     79  97.3  84.5  1358 169.    32.2  <tibble [30 × 2…
#> 3 birth_year     8  35       52  87.6  72     896 155.    50.6  <tibble [20 × 2…

# with a visualisation too - try to limit number of bins
inspect_num(starwars, breaks = 10)

#> # A tibble: 3 x 10
#>   col_name     min    q1 median  mean    q3   max    sd pcnt_na hist            
#>   <chr>      <dbl> <dbl>  <dbl> <dbl> <dbl> <dbl> <dbl>   <dbl> <named list>    
#> 1 height        66 167      180 174.  191     264  34.8    6.90 <tibble [13 × 2…
#> 2 mass          15  55.6     79  97.3  84.5  1358 169.    32.2  <tibble [16 × 2…
#> 3 birth_year     8  35       52  87.6  72     896 155.    50.6  <tibble [11 × 2…

# compare two data frames
inspect_num(starwars, starwars[-c(1:10), ], breaks = 10)

#> # A tibble: 3 x 5
#>   col_name   hist_1            hist_2                 jsd fisher_p
#>   <chr>      <list>            <list>               <dbl>    <dbl>
#> 1 height     <tibble [13 × 2]> <tibble [13 × 2]> 0.00264     1.000
#> 2 mass       <tibble [16 × 2]> <tibble [16 × 2]> 0.000239    1    
#> 3 birth_year <tibble [11 × 2]> <tibble [11 × 2]> 0.00218     0.983