Summary and comparison of column types

For a single dataframe, summarise the column types. If two dataframes are supplied, compare column type composition of both dataframes.

inspect_types(df1, df2 = NULL, show_plot = FALSE)

Arguments

df1
A dataframe.

df2
An optional second dataframe for comparison.

show_plot
(Deprecated) Logical flag indicating whether a plot should be shown. Superseded by the function show_plot() and will be dropped in a future version.

Value

A tibble summarising the count and percentage of different column types for one or a pair of data frames.

Details

For a single dataframe, the tibble returned contains the columns:

type, a character vector containing the column types in df1.

cnt, integer counts of each type.

pcnt, the percentage of all columns with each type.

col_name, the names of columns with each type.

For a pair of dataframes, the tibble returned contains the columns:

type, a character vector containing the column types in df1 and df2.

cnt_1, cnt_2, pair of integer columns containing counts of each type - in each of df1 and df2.

pcnt_1, pcnt_2, pair of columns containing the percentage of columns with each type - the data frame name are appended.

For a grouped dataframe, the tibble returned is as for a single dataframe, but where the first k columns are the grouping columns. There will be as many rows in the result as there are unique combinations of the grouping variables.

See also

Examples

# Load dplyr for starwars data & pipe
library(dplyr)

# Single dataframe summary
inspect_types(starwars)

#> # A tibble: 4 x 4
#>   type        cnt  pcnt col_name 
#>   <chr>     <int> <dbl> <list>   
#> 1 character     7 53.8  <chr [7]>
#> 2 list          3 23.1  <chr [3]>
#> 3 numeric       2 15.4  <chr [2]>
#> 4 integer       1  7.69 <chr [1]>

# Paired dataframe comparison
inspect_types(starwars, starwars[1:20, ])

#> # A tibble: 4 x 5
#>   type      cnt_1 pcnt_1 cnt_2 pcnt_2
#>   <chr>     <int>  <dbl> <int>  <dbl>
#> 1 character     7  53.8      7  53.8 
#> 2 list          3  23.1      3  23.1 
#> 3 numeric       2  15.4      2  15.4 
#> 4 integer       1   7.69     1   7.69