Summarise and compare column types in one or two dataframes.

Summarise and compare column types in one or two dataframes.

inspect_types(df1, df2 = NULL, show_plot = FALSE)

Arguments

df1
A data frame.

df2
An optional second data frame for comparison.

show_plot
(Deprecated) Logical flag indicating whether a plot should be shown. Superseded by the function show_plot() and will be dropped in a future version.

Value

A tibble summarising the count and percentage of different column types for one or a pair of data frames.

Details

When df2 = NULL, a tibble is returned with the columns

type character vector containing the column types in df1.

cnt integer counts of each type.

pcnt the percentage of all columns with each type.

col_name the names of columns with each type.

When a second data frame df2 is specified, column type summaries are tabulated for both data frames to enable comparison of contents. The resulting tibble has the columns

type character vector containing the column types in df1 and df2.

cnt_1, cnt_2 pair of integer columns containing counts of each type - in each of df1 and df2.

pcnt_1, pcnt_2 pair of columns containing the percentage of columns with each type - the data frame name are appended.

Examples

data("starwars", package = "dplyr")
# get tibble of column types for the starwars data
inspect_types(starwars)

#> # A tibble: 4 x 4
#>   type        cnt  pcnt col_name 
#>   <chr>     <int> <dbl> <list>   
#> 1 character     7 53.8  <chr [7]>
#> 2 list          3 23.1  <chr [3]>
#> 3 numeric       2 15.4  <chr [2]>
#> 4 integer       1  7.69 <chr [1]>

# compare two data frames
inspect_types(starwars, starwars[, -1])

#> # A tibble: 4 x 5
#>   type      cnt_1 pcnt_1 cnt_2 pcnt_2
#>   <chr>     <int>  <dbl> <int>  <dbl>
#> 1 character     7  53.8      6  50   
#> 2 list          3  23.1      3  25   
#> 3 numeric       2  15.4      2  16.7 
#> 4 integer       1   7.69     1   8.33